requirements.txt
asgiref==3.2.7
Django==3.0
pytz==2020.1
sqlparse==0.3.1
psycopg2==2.8.4
django-cors-headers==3.1.1
djangorestframework==3.10.3
whitenoise==4.1.4
